task1/requirements.txt
Flask==3.0.3
waitress==3.0.0
beautifulsoup4==4.12.3
lxml==5.2.2
requests==2.32.3
tqdm==4.66.4
whoosh==2.7.4
APScheduler==3.10.4
PyYAML==6.0.2
